requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1
pandas==2.1.4
scikit-learn==1.3.2
numpy==1.24.3
joblib==1.3.2

# Примечание: psycopg2-binary нужен только для PostgreSQL
# Для SQLite дополнительные пакеты не требуются (встроен в Python)

